### 2019-09-08 #1718 @arfon
Thanks for this submission @shreyasbapat. While this looks like a useful piece of code, it feels like it would be much more useful contributed to an existing core package such as [`astropy.fits`](https://docs.astropy.org/en/stable/io/fits/).

Have you considered trying to contribute to this package?
