### 2019-09-08 #1718 @shreyasbapat
Hi @arfon
I think astropy is kind of a general utility package while this is more specific. I can ask in the chatroom but I am very sure that the developers will deny for the same kind of reason .
