### 2019-09-10 #1718 @shreyasbapat
Hi @afron I got the reply:
`If it is not in the official FITS standards, it is not for astropy.io.fits, I am afraid.`
They have suggested to make it a affiliated package.
