### 2019-09-16 #1718 @arfon
@shreyasbapat - are there any folks you would suggest as potential reviewers here?

Also, before sending this out for review, it would be good if you could augment/extend the tests. The current automated tests look to be [very limited in scope](https://github.com/shreyasbapat/python-oifits/tree/master/src/oifits/tests).
