### 2019-09-16 #1718 @shreyasbapat
I added tests.
And as I said earlier,

I am not sure what tags to look for in this case, so I will go with Data Science.
Here are my suggestions for potential reviewers: @jkahn @trallard

Thanks @arfon
