### 2019-09-17 #1743 @whedon
[REVIEW]: oifits :  A Python Package to Analyse EHT Data
**Submitting author:** @shreyasbapat (<a href="http://orcid.org/0000-0002-0870-4665">Shreyas Bapat</a>)
**Repository:** <a href="https://github.com/shreyasbapat/python-oifits" target ="_blank">https://github.com/shreyasbapat/python-oifits</a>
**Version:** v0.2.0
