### 2019-09-17 #1743 @arfon
@taldcroft, @mwcraig - many thanks for agreeing to review this submission!

Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
