### 2019-09-22 shreyasbapat/python-oifits#3 @mwcraig
JOSS Review: Missing API documentation 
Looking through the documentation for the package I noticed that although there is a page for API documentation at https://oifits.readthedocs.io/en/latest/api/index.html there is no API documentation at that page.

Part of https://github.com/openjournals/joss-reviews/issues/1743
