### 2019-09-22 shreyasbapat/python-oifits#4 @mwcraig
JOSS Review: Installation instructions should include list of dependencies
One of the checklist items is "Installation instructions: Is there a clearly-stated list of dependencies?". I don't see that list in the installation instructions.

In addition, the dev install instructions say `einsteinpy` in a couple of places instead of `oifits`.

https://github.com/openjournals/joss-reviews/issues/1743
