### 2019-09-22 shreyasbapat/python-oifits#5 @mwcraig
JOSS Review: Example usage should be more extensive
The example usage at the moment includes a single, very brief, jupyter notebook. Having jupyter notebooks as the primary example documentation is fine (and even a single notebook is fine) but there needs to be more extensive explanation in the notebook. For example, an example that included loading some data from the EHT data site (Assuming there is one) and extracting some physical information from it would be much more helpful.

https://github.com/openjournals/joss-reviews/issues/1743
