### 2019-09-22 #1743 @mwcraig
@arfon my review of this submission is complete for the moment.

I have not reviewed the paper itself in detail because of what is, I believe, a major issue with the code, described at https://github.com/shreyasbapat/python-oifits/issues/2

Because of that issue, my recommendation is that this submission needs major revision before it can be accepted.

The other issues I've opened in the code repo are relatively straightforward to address.
