### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
Right! I am sorry. I just noticed. I actually didn't know about this. I will rectify the issue ASAP.
I included the copyright notice where I thought I was using the code. Sorry for this.
