### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
I actually first got @pboley 's code to work with latest python, then wrote my own classes instead, but somehow, I pushed the older code. I will understand what happened and update the repository ASAP. I included the copyright line in the `dataclasses` module, which was strictly the old code. I am sorry for all this mess. Maybe it's because of my negligence towards it. I am genuinely sorry for this type of mistake.
