### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
I understood what I did wrong. I am sorry. I created the same module in two different folders in my system, where in one I made this code work with my datasets, and then simply created a desired API (which I will want) and in the other one I created a different module by my API. And due to negligence, for which I am really sorry, I pushed the first one (Which I never intended to). And then overtime I never noticed as the APIs were exactly same, and Most of the methods were also same (Beacause I designed them so).
I am really ashamed of this small mistake. I have taken down the code, and I will polish the other modules, test it and push it ASAP. Sorry for this @mwcraig
