### 2019-09-22 shreyasbapat/python-oifits#4 @shreyasbapat
Oh, true. I took that directly from einsteinpy. And again. This is all because of the hurry. I am really really sorry for this kind of submission. I will never repeat these mistakes again.
