### 2019-09-22 shreyasbapat/python-oifits#5 @shreyasbapat
Yes, the test.oifits is actually from EHT. I will add this too.
