### 2019-09-22 shreyasbapat/python-oifits#2 @mwcraig
@shreyasbapat  -- mistakes happen, and the purpose of review is to catch them so that they can be fixed. I tried to keep my original comment as dispassionate as possible to allow the possibility for fixing the issue. I won't have time to come back around to this for a few more days (more likely next weekend), so there is time from my point of view.
