### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
Thanks a lot @mwcraig . This was really unintentional. I will fix this in time.
