### 2019-09-22 shreyasbapat/python-oifits#2 @pboley
Well, I just saw the JOSS draft and review from @mwcraig.  Quickly opened the repo and saw more or less my python oifits module copied verbatim.  Perplexing.
