### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
Yes, true. I am sorry. Never intended to publish this. Sorry @pboley .
