### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
For the sake of the problem that occurred, I will remove the submission. I am again sorry @pboley . I will submit again with a proper submission
