### 2019-09-22 shreyasbapat/python-oifits#2 @pboley
From what I can understand, you created a new module using the same API as an existing module.

What was the motivation?

Yet you write this in the readme:
"oifits is a part of my Major Technical Project of my B.Tech. Degree. I faced the issue of not understanding exactly what the data was, and there was NO PYTHON MODULE which can tell me what's inside the data file. So I spent a significant amount of my research in understanding the data. And I don't want this to happen to anyone who wants to do their research using the world's largest telescope ;)"

Since you apparently wrote a second module, maybe you could just share the repo?
