### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
The motivation was to make the data easily readable for EHT related applications.

The mistake I made was, I used one of your 2013 codes to read the data, but then realised that I wanted the API to be different and easy for my usecase.

I rewrote the code. And I pushed the alpha version I created using your 2013 code (This was by mistake.) I was planning to use your DataClasses where I wrote the copyright notice from your licence. https://github.com/shreyasbapat/python-oifits/blob/master/src/oifits/dataclasses.py#L3
But as all of the code was messed up, it makes actually no sense. I understand the concerns. I have removed the code as soon as I realised. I am really sorry. I will share the code once I make that available online. But due to some exams, I won't be able to do that right now.

Meanwhile https://github.com/openjournals/joss-reviews/issues/1743#issuecomment-533917310
I have asked to remove the submission. I am sorry once again. This was negligence in my part and I had no intent to do so. (Otherwise there was no actual meaning of including your copyright notice in one of the modules) . I am sorry for all the trouble caused.
