### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
I will also remove all the code from the history of git. As it is unjust to have that otherwise. I am very sorry on this part, I realise this is a mistake. I seek your apology for the same. And I hope you will forgive me for this @pboley .
