### 2019-09-22 c1e47e2
Remove all the unjust code
