+++ paper.bib
@@ -1,0 +1,31 @@
+@article{astropy,
+    author = {{Astropy Collaboration}},
+    title = "{Astropy: A community Python package for astronomy}",
+    journal = {\aap},
+    archivePrefix = "arXiv",
+    eprint = {1307.6212},
+    primaryClass = "astro-ph.IM",
+    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+    year = 2013,
+    month = oct,
+    volume = 558,
+    doi = {10.1051/0004-6361/201322068},
+    adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
+}
+
+@book{oliphant2006guide,
+  title={A guide to NumPy},
+  author={Oliphant, Travis E},
+  volume={1},
+  year={2006},
+  publisher={Trelgol Publishing USA}
+}
+
+@article{1510.04556,
+    Author = {Gilles Duvert and John Young and Christian Hummel},
+    Title = {OIFITS 2: the 2nd version of the Data Exchange Standard for Optical (Visible/IR) Interferometry},
+    Year = {2015},
+    Eprint = {arXiv:1510.04556},
+    Howpublished = {A&A 597, A8 (2017)},
+    Doi = {10.1051/0004-6361/201526405},
+}
