+++ docs/source/jupyter.rst
@@ -1,0 +1,7 @@
+Jupyter notebooks
+=================
+
+.. toctree::
+   :maxdepth: 1
+
+   /examples/example.ipynb
