+++ docs/source/api/data.rst
@@ -1,0 +1,5 @@
+Data module
+===========
+
+.. automodule:: oifits.data
+    :members:
