+++ docs/source/api/utils.rst
@@ -1,0 +1,5 @@
+Utils module
+============
+
+.. automodule:: oifits.utils
+    :members:
