+++ docs/source/api/index.rst
@@ -1,0 +1,16 @@
+python-oifits API
+=================
+
+Welcome to the API documenatation of oifits. Please navigate through the given modules to get to
+know the API of the classes and methods. If you find anything missing, please open an `issue in the repo`_ .
+
+.. _`issue in the repo` : https://github.com/shreyasbapat/python-oifits
+
+.. toctree::
+    :hidden:
+    :maxdepth: 2
+
+    data
+    dataclasses
+    read
+    utils

[…]
