### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
Since I cleaned the history, links won't open.
Here is how the dataclass.py header looked:
```
"""
These Classes are inspired by Paul Boley's code for OIFITS Data!
Copyright (c) 2013, Paul Boley
Copyright (c) 2019, Shreyas Bapat
"""

I can definitely not justify this, but this was really unintentional.
```
