### 2019-09-22 shreyasbapat/python-oifits#2 @shreyasbapat
I  have made the code private to https://github.com/shreyasbapat/bad-oifits
It was a wrong push that i made, and I have removed all those file from this repo. Please ask me the rights to view the private repository, in order to understand what actually happened.
