### 2019-09-22 #1743 @arfon
Thanks for flagging these issues @mwcraig. As @shreyasbapat has requested, at this point I think it's best to have them withdraw this submission and potentially resubmit at some point in the future when these issues have been addressed.

> Can you stop the review at this point so that I can correct my mistakes?

Certainly. I'll close this issue now.

Thanks for your efforts on this @mwcraig! @shreyasbapat - if you're able to resolve these issues to the satisfaction of @pboley and the reviewers here we'd welcome a submission at some point again in the future.
